CHALLENGE-1/requirements.txt
torch
torchvision
timm
pandas
scikit-learn
tqdm
Pillow
